Ensure a VIDEO_SERVED event is initiated for each video watched
When a video updates it’s time position we initiate a one minute timer, each new event will cancel the previous - This ensures that an event is sent for any watched video.
Additionally, if a user closes the web browser or exits through navigation (back button unloading the app etc.) we will send off a served video event for all currently in progress videos.